+++ ur_operator/uptimerobot.py
@@ -2,15 +2,16 @@ import logging
 import os
 
 import uptimerobotpy as ur
+import config
 
 
 def create_uptimerobot_api():
     try:
-        ur_api_key = os.environ['UPTIMEROBOT_API_KEY']
+        ur_api_key = config.Config().UPTIMEROBOT_API_KEY
     except KeyError as error:
-        logging.error(
-            'Required environment variable %s has not been provided', error.args[0])
-        raise error
+        msg = f'Required environment variable {error.args[0]} has not been provided'
+        logging.error(msg)
+        raise RuntimeError(msg)
 
     uptime_robot = ur.UptimeRobot(api_key=ur_api_key)
     resp = uptime_robot.get_account_details()