+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- allow to disable creating/updating/deleting monitors for Ingress resources [#6](https://github.com/brennerm/uptimerobot-operator/pull/6)
+
 ## [0.1.0] - 2021-02-02
 
 ### Added